Floats to replace ints for config

+++ config.py
@@ -23,7 +23,7 @@ print (config_path)
 with open('/home/pi/Desktop/VSCADA/config.csv' , newline='') as csvfile:
     reader = csv.reader(csvfile, delimiter=',', quotechar='|')
     for row in reader:
-        sensor = SensorInfo(name=row[0], lower_threshold=int(row[1]), upper_threshold=int(row[2]), drop_out=int(row[3]))
+        sensor = SensorInfo(name=row[0], lower_threshold=float(row[1]), upper_threshold=float(row[2]), drop_out=float(row[3]))
         sensor_thresh_list.append(sensor)
 
 #print (sensor_thresh_list)
